refactor: extract MockUserManagementService and StoragesMockHelper from Simulator
