+++ core/src/test/java/com/zextras/carbonio/files/Simulator.java
@@ -4,8 +4,12 @@
 
 package com.zextras.carbonio.files;
 
-import com.google.inject.*;
+import com.google.inject.AbstractModule;
+import com.google.inject.Guice;
+import com.google.inject.Injector;
 import com.google.inject.Module;
+import com.google.inject.Provides;
+import com.google.inject.Singleton;
 import com.google.inject.util.Modules;
 import com.zextras.carbonio.files.Constants.Config.Database;
 import com.zextras.carbonio.files.Constants.ServiceDiscover.Config.Key;
@@ -16,44 +20,25 @@ import com.zextras.carbonio.files.dal.DatabaseManager;
 import com.zextras.carbonio.files.dal.dao.ebean.Node;
 import com.zextras.carbonio.files.netty.HttpRoutingHandler;
 import com.zextras.carbonio.files.utilities.MockFilesConfig;
-import com.zextras.carbonio.user_management.sdk.grpc.GetUserByEmailRequest;
-import com.zextras.carbonio.user_management.sdk.grpc.GetUserByIdRequest;
-import com.zextras.carbonio.user_management.sdk.grpc.GetUserMyselfRequest;
-import com.zextras.carbonio.user_management.sdk.grpc.UserInfoProto;
-import com.zextras.carbonio.user_management.sdk.grpc.UserInfoResponse;
+import com.zextras.carbonio.files.utilities.MockUserManagementService;
 import com.zextras.carbonio.user_management.sdk.grpc.UserManagementServiceGrpc;
 import com.zextras.carbonio.user_management.sdk.grpc.UserManagementServiceGrpc.UserManagementServiceBlockingStub;
-import com.zextras.carbonio.user_management.sdk.grpc.UserManagementServiceGrpc.UserManagementServiceImplBase;
-import com.zextras.carbonio.user_management.sdk.grpc.UserMyselfProto;
-import com.zextras.carbonio.user_management.sdk.grpc.UserMyselfResponse;
-import com.zextras.carbonio.user_management.sdk.grpc.UserTypeProto;
-import com.zextras.storages.internal.pojo.Query;
-import com.zextras.storages.internal.pojo.StoragesBulkDeleteResponse;
 import io.grpc.ManagedChannel;
 import io.grpc.Server;
-import io.grpc.Status;
 import io.grpc.inprocess.InProcessChannelBuilder;
 import io.grpc.inprocess.InProcessServerBuilder;
-import io.grpc.stub.StreamObserver;
 import io.netty.channel.embedded.EmbeddedChannel;
 import io.netty.handler.codec.http.HttpMethod;
-
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
-
 import org.mockserver.client.MockServerClient;
 import org.mockserver.integration.ClientAndServer;
 import org.mockserver.model.HttpRequest;
 import org.mockserver.model.HttpResponse;
-import org.mockserver.model.JsonBody;
-import org.mockserver.model.Parameter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testcontainers.containers.PostgreSQLContainer;
@@ -299,14 +284,6 @@ public class Simulator implements AutoCloseable {
     return this;
   }
 
-  /**
-   * Registers a token-to-userId mapping so that subsequent gRPC {@code getUserMyself} calls
-   * with the given token will return a valid user with the "carbonioFeatureFilesEnabled" feature.
-   */
-  private void registerUser(String cookie, String userId) {
-    mockUmService.registerToken(cookie, userId);
-  }
-
   private Simulator startStorages() {
     startMockServer();
 
@@ -435,15 +412,11 @@ public class Simulator implements AutoCloseable {
     return injector;
   }
 
-  public MockServerClient getServiceDiscoverMock() {
-    return serviceDiscoverMock;
-  }
-
   /**
    * Returns the mock UM gRPC service, allowing tests to register additional users
    * (e.g. for getUserById lookups in transfer ownership scenarios).
    */
-  public MockUserManagementService getUserManagementService() {
+  public MockUserManagementService getUserManagementMock() {
     return mockUmService;
   }
 
@@ -473,11 +446,11 @@ public class Simulator implements AutoCloseable {
     return storagesMock;
   }
 
-  public MockServerClient getPreviewServiceMock() {
+  public MockServerClient getPreviewMock() {
     return previewServiceMock;
   }
 
-  public MockServerClient getDocsConnectorServiceMock() {
+  public MockServerClient getDocsConnectorMock() {
     return docsConnectorServiceMock;
   }
 
@@ -516,40 +489,6 @@ public class Simulator implements AutoCloseable {
     }
   }
 
-  public void getBlob(String nodeId, int version) {
-    storagesMock
-        .when(
-            HttpRequest.request()
-                .withMethod(HttpMethod.GET.toString())
-                .withPath("/download")
-                .withQueryStringParameter(Parameter.param("node", nodeId))
-                .withQueryStringParameter(Parameter.param("version", String.valueOf(version)))
-                .withQueryStringParameter(Parameter.param("type", "files")))
-        .respond(
-            HttpResponse.response()
-                .withStatusCode(200)
-                .withBody((nodeId + version).getBytes(StandardCharsets.UTF_8)));
-  }
-
-  public void bulkDelete(List<String> ids) {
-    final StoragesBulkDeleteResponse response = new StoragesBulkDeleteResponse();
-    Query queryList = new Query();
-    for (String id : ids) {
-      queryList.setNode(id);
-      queryList.setType("files");
-    }
-    response.setIds(List.of(queryList));
-    storagesMock
-        .when(
-            HttpRequest.request()
-                .withMethod(HttpMethod.POST.toString())
-                .withPath("/bulk-delete"))
-        .respond(
-            HttpResponse.response()
-                .withStatusCode(200)
-                .withBody(JsonBody.json(response)));
-  }
-
   /**
    * Guice module that overrides the ManagedChannel and BlockingStub bindings
    * to use the gRPC InProcess transport for testing.
@@ -572,117 +511,6 @@ public class Simulator implements AutoCloseable {
     @Singleton
     public UserManagementServiceBlockingStub provideUserManagementStub() {
       return UserManagementServiceGrpc.newBlockingStub(channel);
-    }
-  }
-
-  /**
-   * In-memory gRPC service implementation for UserManagement. Supports
-   * {@code getUserMyself} (token lookup), {@code getUserById} (userId lookup),
-   * and {@code getUserByEmail} (email lookup).
-   */
-  public static class MockUserManagementService extends UserManagementServiceImplBase {
-
-    private final Map<String, UserMyselfResponse> tokenToMyself = new ConcurrentHashMap<>();
-    private final Map<String, UserInfoProto> userIdToInfo = new ConcurrentHashMap<>();
-
-    /**
-     * Registers a minimal token-to-userId mapping. A full {@link UserMyselfResponse} is built
-     * with default values for email, name, domain, status, locale, and the
-     * "carbonioFeatureFilesEnabled" feature enabled.
-     */
-    void registerToken(String token, String userId) {
-      if (!tokenToMyself.containsKey(token)) {
-        UserInfoProto info = UserInfoProto.newBuilder()
-            .setUserId(userId)
-            .setEmail("fake-email@example.com")
-            .setFullName("Fake User")
-            .setDomain("example.com")
-            .setStatus("active")
-            .setType(UserTypeProto.INTERNAL)
-            .build();
-        UserMyselfProto myself = UserMyselfProto.newBuilder()
-            .setInfo(info)
-            .setLocale("en")
-            .addFeatures("carbonioFeatureFilesEnabled")
-            .build();
-        tokenToMyself.put(token, UserMyselfResponse.newBuilder().setUser(myself).build());
-        userIdToInfo.put(userId, info);
-      }
-    }
-
-    /**
-     * Removes a userId from the {@code getUserById} lookup map so that subsequent
-     * {@code getUserById} calls for this user will return NOT_FOUND.
-     */
-    public void unregisterUserById(String userId) {
-      userIdToInfo.remove(userId);
-    }
-
-    /**
-     * Registers a user profile for lookup by userId via {@code getUserById}.
-     * This is used by integration tests that need to look up users other than
-     * the requester (e.g. transfer ownership target user).
-     */
-    public void registerUserById(String userId, String email, String fullName,
-        String domain, String status) {
-      UserInfoProto info = UserInfoProto.newBuilder()
-          .setUserId(userId)
-          .setEmail(email)
-          .setFullName(fullName)
-          .setDomain(domain)
-          .setStatus(status)
-          .setType(UserTypeProto.INTERNAL)
-          .build();
-      userIdToInfo.put(userId, info);
-    }
-
-    void clearAll() {
-      tokenToMyself.clear();
-      userIdToInfo.clear();
-    }
-
-    @Override
-    public void getUserMyself(GetUserMyselfRequest request,
-        StreamObserver<UserMyselfResponse> responseObserver) {
-      String token = request.getToken();
-      UserMyselfResponse response = tokenToMyself.get(token);
-      if (response != null) {
-        responseObserver.onNext(response);
-        responseObserver.onCompleted();
-      } else {
-        responseObserver.onError(
-            Status.UNAUTHENTICATED.withDescription("Invalid token").asRuntimeException());
-      }
-    }
-
-    @Override
-    public void getUserById(GetUserByIdRequest request,
-        StreamObserver<UserInfoResponse> responseObserver) {
-      String userId = request.getUserId();
-      UserInfoProto info = userIdToInfo.get(userId);
-      if (info != null) {
-        responseObserver.onNext(UserInfoResponse.newBuilder().setUser(info).build());
-        responseObserver.onCompleted();
-      } else {
-        responseObserver.onError(
-            Status.NOT_FOUND.withDescription("User not found").asRuntimeException());
-      }
-    }
-
-    @Override
-    public void getUserByEmail(GetUserByEmailRequest request,
-        StreamObserver<UserInfoResponse> responseObserver) {
-      String email = request.getUserEmail();
-      // Search by email across registered users
-      for (UserInfoProto info : userIdToInfo.values()) {
-        if (info.getEmail().equals(email)) {
-          responseObserver.onNext(UserInfoResponse.newBuilder().setUser(info).build());
-          responseObserver.onCompleted();
-          return;
-        }
-      }
-      responseObserver.onError(
-          Status.NOT_FOUND.withDescription("User not found by email").asRuntimeException());
     }
   }
 
@@ -716,10 +544,7 @@ public class Simulator implements AutoCloseable {
 
     public SimulatorBuilder withUserManagement(Map<String, String> users) {
       simulator.startUserManagement();
-      users.forEach(
-          (cookie, userId) -> {
-            simulator.registerUser(cookie, userId);
-          });
+      users.forEach((cookie, userId) -> simulator.mockUmService.registerToken(cookie, userId));
       return this;
     }
