+++ core/src/test/java/com/zextras/carbonio/files/api/DeleteAllNodesAndBlobsApiIT.java
@@ -15,6 +15,7 @@ import com.zextras.carbonio.files.api.utilities.entities.SimplePopulatorTextFile
 import com.zextras.carbonio.files.dal.repositories.interfaces.FileVersionRepository;
 import com.zextras.carbonio.files.dal.repositories.interfaces.LinkRepository;
 import com.zextras.carbonio.files.dal.repositories.interfaces.NodeRepository;
+import com.zextras.carbonio.files.utilities.StoragesMockHelper;
 import com.zextras.carbonio.files.utilities.http.HttpRequest;
 import com.zextras.carbonio.files.utilities.http.HttpResponse;
 import org.assertj.core.api.Assertions;
@@ -30,6 +31,7 @@ import java.util.Optional;
 class DeleteAllNodesAndBlobsApiIT {
 
   static Simulator simulator;
+  static StoragesMockHelper storagesMockHelper;
   static NodeRepository nodeRepository;
   static FileVersionRepository fileVersionRepository;
   static LinkRepository linkRepository;
@@ -53,6 +55,7 @@ class DeleteAllNodesAndBlobsApiIT {
     nodeRepository = injector.getInstance(NodeRepository.class);
     fileVersionRepository = injector.getInstance(FileVersionRepository.class);
     linkRepository = injector.getInstance(LinkRepository.class);
+    storagesMockHelper = new StoragesMockHelper(simulator.getStoragesMock());
   }
 
   @AfterEach
@@ -123,7 +126,7 @@ class DeleteAllNodesAndBlobsApiIT {
             .withWantedResultFormat("")
             .build();
 
-    simulator.bulkDelete(List.of(new String[]{"00000000-0000-0000-0000-000000000002", "00000000-0000-0000-0000-000000000003"}));
+    storagesMockHelper.bulkDelete(List.of(new String[]{"00000000-0000-0000-0000-000000000002", "00000000-0000-0000-0000-000000000003"}));
 
     List<Map.Entry<String, String>> headers = List.of(Map.entry("Internal", ""));
     final HttpRequest httpRequest =
