+++ core/src/test/java/com/zextras/carbonio/files/api/DownloadByPublicLinkApiIT.java
@@ -14,6 +14,7 @@ import com.zextras.carbonio.files.dal.dao.ebean.NodeType;
 import com.zextras.carbonio.files.dal.repositories.interfaces.FileVersionRepository;
 import com.zextras.carbonio.files.dal.repositories.interfaces.LinkRepository;
 import com.zextras.carbonio.files.dal.repositories.interfaces.NodeRepository;
+import com.zextras.carbonio.files.utilities.StoragesMockHelper;
 import com.zextras.carbonio.files.utilities.http.HttpRequest;
 import com.zextras.carbonio.files.utilities.http.HttpResponse;
 import io.netty.handler.codec.http.HttpMethod;
@@ -33,6 +34,7 @@ import org.mockserver.verify.VerificationTimes;
 public class DownloadByPublicLinkApiIT {
 
   static Simulator simulator;
+  static StoragesMockHelper storagesMockHelper;
   static NodeRepository nodeRepository;
   static FileVersionRepository fileVersionRepository;
   static LinkRepository linkRepository;
@@ -53,6 +55,7 @@ public class DownloadByPublicLinkApiIT {
     nodeRepository = injector.getInstance(NodeRepository.class);
     fileVersionRepository = injector.getInstance(FileVersionRepository.class);
     linkRepository = injector.getInstance(LinkRepository.class);
+    storagesMockHelper = new StoragesMockHelper(simulator.getStoragesMock());
   }
 
   @AfterEach
@@ -102,7 +105,7 @@ public class DownloadByPublicLinkApiIT {
             Optional.empty(),
             Optional.empty());
 
-    simulator.getBlob("00000000-0000-0000-0000-000000000000", 1);
+    storagesMockHelper.getBlob("00000000-0000-0000-0000-000000000000", 1);
 
     final String publicLinkUrl = publicLinkEndpoint + publicLinkId;
     final HttpRequest httpRequest = HttpRequest.of("GET", publicLinkUrl, userToken, null);
@@ -163,7 +166,7 @@ public class DownloadByPublicLinkApiIT {
             Optional.empty(),
             Optional.of("test"));
 
-    simulator.getBlob("00000000-0000-0000-0000-000000000000", 1);
+    storagesMockHelper.getBlob("00000000-0000-0000-0000-000000000000", 1);
 
     final String publicLinkUrl = publicLinkEndpoint + publicLinkId;
     final HttpRequest httpRequest = HttpRequest.of("GET", publicLinkUrl, userToken, null);
@@ -372,7 +375,7 @@ public class DownloadByPublicLinkApiIT {
             Optional.empty())
         .addNodeToTrash("00000000-0000-0000-0000-000000000000", "LOCAL_ROOT");
 
-    simulator.getBlob("00000000-0000-0000-0000-000000000000", 1);
+    storagesMockHelper.getBlob("00000000-0000-0000-0000-000000000000", 1);
 
     final String publicLinkUrl = publicLinkEndpoint + publicLinkId;
     final HttpRequest httpRequest = HttpRequest.of("GET", publicLinkUrl, userToken, null);
