+++ core/src/test/java/com/zextras/carbonio/files/utilities/StoragesMockHelper.java
@@ -1,0 +1,59 @@
+// SPDX-FileCopyrightText: 2026 Zextras <https://www.zextras.com>
+//
+// SPDX-License-Identifier: AGPL-3.0-only
+
+package com.zextras.carbonio.files.utilities;
+
+import com.zextras.storages.internal.pojo.Query;
+import com.zextras.storages.internal.pojo.StoragesBulkDeleteResponse;
+import io.netty.handler.codec.http.HttpMethod;
+import java.nio.charset.StandardCharsets;
+import java.util.List;
+import org.mockserver.client.MockServerClient;
+import org.mockserver.model.HttpRequest;
+import org.mockserver.model.HttpResponse;
+import org.mockserver.model.JsonBody;
+import org.mockserver.model.Parameter;
+
+public class StoragesMockHelper {
+
+  private final MockServerClient storagesMock;
+
+  public StoragesMockHelper(MockServerClient storagesMock) {
+    this.storagesMock = storagesMock;
+  }
+
+  public void getBlob(String nodeId, int version) {
+    storagesMock
+        .when(
+            HttpRequest.request()
+                .withMethod(HttpMethod.GET.toString())
+                .withPath("/download")
+                .withQueryStringParameter(Parameter.param("node", nodeId))
+                .withQueryStringParameter(Parameter.param("version", String.valueOf(version)))
+                .withQueryStringParameter(Parameter.param("type", "files")))
+        .respond(
+            HttpResponse.response()
+                .withStatusCode(200)
+                .withBody((nodeId + version).getBytes(StandardCharsets.UTF_8)));
+  }
+
+  public void bulkDelete(List<String> ids) {
+    final StoragesBulkDeleteResponse response = new StoragesBulkDeleteResponse();
+    Query queryList = new Query();
+    for (String id : ids) {
+      queryList.setNode(id);
+      queryList.setType("files");
+    }
+    response.setIds(List.of(queryList));
+    storagesMock
+        .when(
+            HttpRequest.request()
+                .withMethod(HttpMethod.POST.toString())
+                .withPath("/bulk-delete"))
+        .respond(
+            HttpResponse.response()
+                .withStatusCode(200)
+                .withBody(JsonBody.json(response)));
+  }
+}
